requirements.txt
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Script==2.0.6
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
python-dotenv==0.12.0
six==1.14.0
Werkzeug==1.0.0
